n/a

+++ MyProject/stage 04/src/main/java/com/example/myproject/web/OffersController.java
@@ -3,13 +3,13 @@ package com.example.myproject.web;
 import com.example.myproject.model.binding.OfferAddBindingModel;
 import com.example.myproject.model.view.OfferSummaryView;
 import com.example.myproject.service.OffersService;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
@@ -40,6 +40,7 @@ public class OffersController {
                            BindingResult bindingResult,
                            RedirectAttributes redirectAttributes,
                            Principal principal) {
+        System.out.println();
 
         if (bindingResult.hasErrors()) {
             redirectAttributes
@@ -72,10 +73,25 @@ public class OffersController {
         return "offer-remove";
     }
 
+
+//    @PreAuthorize("@offersServiceImpl.isOfferOwner(#principal, #id)")
     @PostMapping("/offers/remove/{id}")
-    public String removeOffer(@PathVariable Long id) {
+    public String removeOffer(@PathVariable Long id, Principal principal) {
+        System.out.println();
+        boolean offerOwnerOrAdmin = this.offersService.isOfferOwner(principal, id);
+        if (!offerOwnerOrAdmin) {
+            throw new UserNotSupportedOperation(principal.getName());
+        }
         this.offersService.removeOffer(id);
         return "redirect:/remove-offer";
+    }
+
+    @ExceptionHandler(UserNotSupportedOperation.class)
+    public ModelAndView handleDbExceptions(UserNotSupportedOperation e) {
+        ModelAndView modelAndView = new ModelAndView("userNotSupportedOperation");
+        modelAndView.addObject("userName", e.getUsername());
+        modelAndView.setStatus(HttpStatus.UNAUTHORIZED);
+        return modelAndView;
     }
